pyproject.toml
[project]
name = "tsimcne"
version = "0.0.1"
description = "Contrastive experiments"
authors = [
    {name = "Niklas Böhm", email = "jan-niklas.boehm@uni-tuebingen.de"},
]
dependencies = [
    "torch",
    "torchvision",
    "pytorch-lightning",
    "numpy",
    "pandas",
    "matplotlib",
    "tqdm",
    "python-telegram-bot",
    "ffmpeg-python",
    "scikit-learn",
    "hdbscan",
    "pillow",
    "annoy",
    "scipy",
    "medmnist",
]
requires-python = ">=3.9"
license = {text = "EUPL"}

[project.urls]
homepage = ""


[tool]
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
